requirements.txt
addict
numpy
openmim
pandas
Pillow
PyYAML
requests
six
timm>=0.9.0
torch>=2.0.0
torchvision
tqdm
transformers>=4.30.0
opencv-python
wandb
pyarrow==13.0.0
tabulate
flask
flask-cors
